trading-data/requirements.txt
numpy
pandas
yahooquery
pyodbc
sqlalchemy
python-dotenv
